Print usage instead of panic for no arguments
Previously, if no arguments were passed to the
root command, a panic was triggered which in turn
bypassed the error handling strategy. Also the
provided message was not really helpful to the
user.

With this commit the usage is printed instead.
Additionally, an error is returned which is
correctly handled by the error handling strategy.